+++ chat-ws/src/main/java/com/tak/chatws/ChatWsApplication.java
@@ -1,13 +1,13 @@
-package Spring_Websocket.websocket;
+package com.tak.chatws;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class WebsocketApplication {
+public class ChatWsApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(WebsocketApplication.class, args);
+		SpringApplication.run(ChatWsApplication.class, args);
 	}
 
 }
