+++ chat-ws/src/main/java/com/tak/chatws/redis/RedisSubscriber.java
@@ -1,6 +1,6 @@
-package Spring_Websocket.websocket.redis;
+package com.tak.chatws.redis;
 
-import Spring_Websocket.websocket.chat.ChatMessageDto;
+import com.tak.chatws.chat.ChatMessageDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -8,8 +8,6 @@ import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Component;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 @Slf4j
 @Component
