+++ chat-ws/src/main/java/com/tak/chatws/websocket/WebSocketSessionLogger.java
@@ -1,10 +1,8 @@
 // src/main/java/Spring_Websocket/websocket/WebSocketSessionLogger.java
-package Spring_Websocket.websocket.websocket;
+package com.tak.chatws.websocket;
 
 import org.springframework.context.event.EventListener;
-import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
-import org.springframework.web.socket.messaging.SessionConnectEvent;
 import org.springframework.web.socket.messaging.SessionConnectedEvent;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
