+++ chat-ws/src/test/java/com/tak/chatws/chat/ChatMessageDtoTest.java
@@ -1,4 +1,4 @@
-package Spring_Websocket.websocket.chat;
+package com.tak.chatws.chat;
 
 import org.junit.jupiter.api.Test;
